requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.35
psycopg2-binary==2.9.11
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
google-generativeai==0.8.3
httpx==0.27.2
pydantic==2.9.2
email-validator==2.2.0
python-multipart==0.0.9
bcrypt==4.0.1